* Fixed failing race and short tests
